only include release note in RPM when one exists

+++ tools/packaging/linux/create_rpm.sh
@@ -45,7 +45,10 @@ echo "Copying application files"
 # License, version, release note...
 cp LICENSE "${AOC_ROOT}/opt/aws/aws-otel-collector/"
 cp VERSION "${AOC_ROOT}/opt/aws/aws-otel-collector/bin/"
-cp "docs/releases/${VERSION}.md" "${AOC_ROOT}/opt/aws/aws-otel-collector/RELEASE_NOTE"
+if [ -s "docs/releases/${VERSION}.md" ]
+  then
+    cp "docs/releases/${VERSION}.md" "${AOC_ROOT}/opt/aws/aws-otel-collector/RELEASE_NOTE"
+fi
 
 # binary
 cp "build/linux/${SOURCE_ARCH}/aoc" "${AOC_ROOT}/opt/aws/aws-otel-collector/bin/aws-otel-collector"
